### 2025-11-16 #9421 @editorialbot
[PRE REVIEW]: Open Computational Chemistry (OCC) - A portable software library and program for quantum chemistry and crystallography
**Submitting author:** @peterspackman (<a href="http://orcid.org/0000-0002-6532-8571">Peter R. Spackman</a>)
**Repository:** https://github.com/peterspackman/occ
**Branch with paper.md** (empty if default branch): paper-joss
**Version:** v0.8.5
