### 2025-11-22 05d1afb
added ccache and lld checks
