### 2025-11-22 357a0a4
moved PES::solve_linear_system to .cpp to compile once
