+++ src/elastic_fit/pes.cpp
@@ -74,6 +74,26 @@ inline void print_vector(const occ::Vec &vec, int per_line) {
   }
 }
 
+occ::Mat PES::solve_linear_system(const occ::Mat &A, const occ::Mat &B,
+                                  LinearSolverType solver_type,
+                                  double svd_threshold) {
+  switch (solver_type) {
+  case LinearSolverType::LU:
+    return A.lu().solve(B);
+  case LinearSolverType::SVD: {
+    Eigen::JacobiSVD<occ::Mat> svd(A, Eigen::ComputeThinU | Eigen::ComputeThinV);
+    svd.setThreshold(svd_threshold);
+    return svd.solve(B);
+  }
+  case LinearSolverType::QR:
+    return A.householderQr().solve(B);
+  case LinearSolverType::LDLT:
+    return A.ldlt().solve(B);
+  default:
+    throw std::runtime_error("Unknown linear solver type");
+  }
+}
+
 occ::Mat6 PES::compute_elastic_tensor(double volume,
                                       LinearSolverType solver_type,
                                       double svd_threshold,
