### 2025-11-27 92cea18
Fixed func from previous branch
Fixed func from previous branch
