+++ include/occ/io/occ_input.h
@@ -112,7 +112,7 @@ struct RuntimeInput {
 };
 
 struct OptimizationInput {
-  // Convergence criteria 
+  // Convergence criteria
   double gradient_max{0.15e-3};     // Maximum gradient component (Ha/Angstrom) - more strict, ~ORCA's 3e-4 Ha/bohr
   double gradient_rms{0.05e-3};     // RMS gradient (Ha/Angstrom) - more strict, ~ORCA's 1e-4 Ha/bohr
   double step_max{1.8e-3};          // Maximum step (Angstrom)
@@ -120,15 +120,15 @@ struct OptimizationInput {
   double energy_change{1e-6};       // Energy convergence (Hartree)
   bool use_energy_criterion{false}; // Whether to use energy criterion
   int max_iterations{100};          // Maximum optimization steps
-  
+
   // Integral accuracy control
   double gradient_integral_precision{1e-10}; // Final gradient integral cutoff
   double early_gradient_integral_precision{1e-8}; // Looser cutoff for early steps
   double tight_gradient_threshold{1e-3}; // Switch to tight precision when |ΔE| < this (Hartree)
-  
+
   // Output options
   bool write_wavefunction_steps{false}; // Write wavefunction at each optimization step
-  
+
   // Frequency analysis
   bool compute_frequencies{false};  // Compute vibrational frequencies after optimization
 };
@@ -189,5 +189,6 @@ template <typename T> OccInput build(const std::string &filename) {
 template <typename T> OccInput build(std::istream &file) {
   return T(file).as_occ_input();
 }
+OccInput read_occ_input_file(const std::string path);
 
 } // namespace occ::io
