+++ src/io/occ_input.cpp
@@ -1,8 +1,11 @@
-
+#include <fstream>
+#include <occ/core/atom.h>
+#include <occ/core/units.h>
 #include <occ/io/occ_input.h>
+#include <filesystem>
+#include <toml++/toml.hpp>
 
 namespace occ::io {
-
 occ::core::Molecule GeometryInput::molecule() const {
   return occ::core::Molecule(elements, positions);
 }
@@ -15,6 +18,83 @@ void GeometryInput::set_molecule(const occ::core::Molecule &mol) {
   for (size_t i = 0; i < elements.size(); i++) {
     positions.push_back(std::array<double, 3>{pos(0, i), pos(1, i), pos(2, i)});
   }
+}
+
+OccInput read_occ_input_file(const std::string path) {
+  // for now only SCF works.
+  OccInput result;
+  auto config = toml::parse_file(path);
+  auto scf = config["scf"];
+  if ( config["cg"] || config["cube"]|| config["describe"]|| config["dimers"]||
+     config["dma"]|| config["elastic"]|| config["elat"]|| config["elastic_fit"]||
+     config["embed"]|| config["iso"]|| config["pair"]|| config["cuts"])
+    throw std::runtime_error("Not yet implemented, only SCF has direct reading implemented.");
+
+  result.runtime.threads = config["threads"].value_or(result.runtime.threads);
+  result.filename = config["input"].value_or(result.filename);
+  result.method.name = scf["method"].value_or(result.method.name);
+  result.basis.name = scf["basis"].value_or(result.basis.name);
+  result.electronic.charge = scf["charge"].value_or(result.electronic.charge);
+  if (auto arr = scf["output"].as_array()) {
+    result.output.formats.clear();
+    result.output.formats.reserve(arr->size());
+
+    for (auto&& elem : *arr) {
+        if (auto val = elem.value<std::string>()) {
+            result.output.formats.push_back(*val);
+        }
+    }
+  }
+  result.electronic.multiplicity = scf["multiplicity"].value_or(result.electronic.multiplicity);
+  result.electronic.spinorbital_kind =  scf["unrestricted"].value_or(result.electronic.spinorbital_kind) ? SpinorbitalKind::Unrestricted : SpinorbitalKind::Restricted;
+  result.driver.driver = scf["driver"].value_or(result.driver.driver);
+  result.basis.basis_set_directory = scf["basis_set_directory"].value_or(result.basis.basis_set_directory);
+  result.method.integral_precision = scf["integral_precision"].value_or(result.method.integral_precision);
+  result.method.use_direct_df_kernels = scf["use_direct_df_kernels"].value_or(result.method.use_direct_df_kernels);
+
+  // DFT grid settings
+  result.method.dft_grid.max_angular_points = scf["dft_grid_max_angular"].value_or(result.method.dft_grid.max_angular_points);
+  result.method.dft_grid.min_angular_points = scf["dft_grid_min_angular"].value_or(result.method.dft_grid.min_angular_points);
+  result.method.dft_grid.radial_precision = scf["dft_grid_radial_precision"].value_or(result.method.dft_grid.radial_precision);
+  result.method.dft_grid.reduced_first_row_element_grid = scf["dft_grid_reduce_light_elements"].value_or(result.method.dft_grid.reduced_first_row_element_grid);
+
+  // basis set
+  result.basis.df_name = scf["df-basis"].value_or(result.basis.df_name);
+  result.basis.ri_basis = scf["ri-basis"].value_or(result.basis.ri_basis);
+  result.basis.spherical = scf["spherical"].value_or(result.basis.spherical);
+
+  result.method.orbital_smearing_sigma = scf["orbital_smearing_sigma"].value_or(result.method.orbital_smearing_sigma);
+
[…]
