+++ 3rdparty/CMakeLists.txt
@@ -14,10 +14,34 @@ cpmaddpackage(
     VERSION "1.x"
 )
 
-cpmaddpackage(
+CPMAddPackage(
     NAME tomlplusplus
     GITHUB_REPOSITORY "marzer/tomlplusplus"
     VERSION "3.4.0"
+)
+
+if(NOT "${USE_SYSTEM_TBB}")
+
+set(TBB_OPTIONS "")
+list(APPEND TBB_OPTIONS "TBB_TEST OFF")
+list(APPEND TBB_OPTIONS "TBB_STRICT OFF")
+# Enable HWLOC for better NUMA topology detection on native builds, disable for Emscripten
+if(EMSCRIPTEN)
+    list(APPEND TBB_OPTIONS "TBB_DISABLE_HWLOC_AUTOMATIC_SEARCH ON")
+else()
+    list(APPEND TBB_OPTIONS "TBB_DISABLE_HWLOC_AUTOMATIC_SEARCH OFF")
+endif()
+
+# Windows MinGW-specific workarounds for TBB compilation issues
+if(WIN32 AND CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
+    list(APPEND TBB_OPTIONS "CMAKE_CXX_FLAGS=-O2 -fno-stack-clash-protection -fno-fcf-protection")
+endif()
+
+CPMAddPackage(
+    NAME oneTBB
+    GITHUB_REPOSITORY "uxlfoundation/oneTBB"
+    VERSION "2022.2.0"
+    OPTIONS ${TBB_OPTIONS}
 )
 
 if(NOT "${USE_SYSTEM_TBB}")
