### 2025-11-28 6db91c8
Fixed filename issue and added verbosity
