### 2025-11-28 9d24ad7
Added run_scf_external for running scf from other programs
