### 2025-12-03 ac2fcfa
Migrate elastic fit to use its own pairs format
