### 2025-12-03 d4c53dc
Bump version number, fix github workflows
