+++ src/occpy.cpp
@@ -57,6 +57,6 @@ NB_MODULE(_occpy, m) {
 #ifdef VERSION_INFO
   m.attr("__version__") = MACRO_STRINGIFY(VERSION_INFO);
 #else
-  m.attr("__version__") = "0.8.7";
+  m.attr("__version__") = "0.8.8";
 #endif
 }
