+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "occpy"
-version = "0.8.7"
+version = "0.8.8"
 description = "A library for quantum chemistry"
 readme = "README.md"
 requires-python = ">=3.10"
