+++ .github/workflows/build_wheels.yml
@@ -32,6 +32,9 @@ jobs:
           - os: ubuntu-latest
             pyver: cp313
             arch: x86_64
+          - os: ubuntu-latest
+            pyver: cp314
+            arch: x86_64
           # macOS ARM64 builds
           - os: macos-latest
             pyver: cp310
@@ -44,6 +47,9 @@ jobs:
             arch: arm64
           - os: macos-latest
             pyver: cp313
+            arch: arm64
+          - os: macos-latest
+            pyver: cp314
             arch: arm64
           # macOS Universal2 builds
           - os: macos-latest
@@ -58,6 +64,9 @@ jobs:
           - os: macos-latest
             pyver: cp313
             arch: universal2
+          - os: macos-latest
+            pyver: cp314
+            arch: universal2
 
     steps:
       - name: Checkout repos
@@ -67,6 +76,9 @@ jobs:
 
       - name: Install uv
         uses: astral-sh/setup-uv@v6
+
+      - name: Clean wheelhouse
+        run: rm -rf wheelhouse
 
       - name: Build wheels
         uses: pypa/cibuildwheel@v3.1.4
@@ -81,6 +93,7 @@ jobs:
         with:
           path: ./wheelhouse/*.whl
           name: occpy-${{ matrix.os }}-${{ matrix.pyver }}-${{ matrix.arch }}
+          overwrite: true
 
   build_pyodide_wheel:
     name: Build wheel for pyodide
@@ -88,6 +101,8 @@ jobs:
     continue-on-error: true  # Allow this job to fail without failing the workflow
     steps:
       - uses: actions/checkout@v5
+      - name: Clean wheelhouse
+        run: rm -rf wheelhouse
       - uses: pypa/cibuildwheel@v3.1.4
         env:
            CIBW_PLATFORM: pyodide
@@ -97,6 +112,7 @@ jobs:
         with:
           path: ./wheelhouse/*.whl
           name: wasm-occpy-pyodide
+          overwrite: true
 
   upload_all:
     runs-on: ubuntu-latest
