+++ dir_4f24fc6e55ee85dc7b7f34360214d5ab.js
@@ -1,5 +1,6 @@
 var dir_4f24fc6e55ee85dc7b7f34360214d5ab =
 [
+    [ "elastic_fit_json.h", "elastic__fit__json_8h.html", "elastic__fit__json_8h" ],
     [ "elastic_fitting.h", "elastic__fitting_8h.html", "elastic__fitting_8h" ],
     [ "monkhorst_pack.h", "monkhorst__pack_8h.html", "monkhorst__pack_8h" ],
     [ "pes.h", "pes_8h.html", "pes_8h" ],

[…]
