### 2025-12-03 458549a
Update workflows, simd + new emsdk for wasm, pyodide build disables exceptions
