+++ pyproject.toml
@@ -48,3 +48,7 @@ archs = ["auto64"]          # Only target 64 bit architectures
 # Needed for full C++17 support on macOS
 [tool.cibuildwheel.macos.environment]
 MACOSX_DEPLOYMENT_TARGET = "10.15"
+
+# Pyodide doesn't support C++ exceptions - disable them and abort on throw
+[tool.cibuildwheel.pyodide.environment]
+CMAKE_ARGS = "-DCMAKE_CXX_FLAGS=-fno-exceptions"
