### 2025-12-03 f17d608
Adjust python workflow
