### 2025-12-03 d38e54e
Adjust node versions for packaging
