+++ .github/workflows/build_nodejs.yml
@@ -43,7 +43,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: [20, 22]
+        node-version: [20, 22, 24]
       fail-fast: false
 
     steps:
@@ -158,7 +158,7 @@ jobs:
     - name: Setup Node.js
       uses: actions/setup-node@v4
       with:
-        node-version: 20
+        node-version: 24
         registry-url: 'https://registry.npmjs.org'
 
     - name: Download WASM build artifacts
@@ -208,7 +208,7 @@ jobs:
     - name: Setup Node.js
       uses: actions/setup-node@v4
       with:
-        node-version: 20
+        node-version: 24
         registry-url: 'https://npm.pkg.github.com'
         scope: '@peterspackman'
 
