+++ src/io/periodic_grid.cpp
@@ -3,12 +3,15 @@
 #include <occ/core/log.h>
 #include <occ/io/cube.h>
 #include <occ/io/periodic_grid.h>
-
+#include <util.h>
+#ifndef M_2_SQRTPI
+#define M_2_SQRTPI 1.1283791670955125585606993
+#endif
 namespace occ::io {
 
 PeriodicGrid::PeriodicGrid() : basis(Mat3::Identity()), m_grid(11, 11, 11) {}
 
-PeriodicGrid::PeriodicGrid(const PeriodicGrid &other) 
+PeriodicGrid::PeriodicGrid(const PeriodicGrid &other)
     : m_grid(other.dimensions()(0), other.dimensions()(1), other.dimensions()(2)),
       title(other.title),
       format(other.format),
@@ -27,7 +30,7 @@ PeriodicGrid &PeriodicGrid::operator=(const PeriodicGrid &other) {
         auto dims = other.dimensions();
         m_grid = geometry::VolumeGrid(dims(0), dims(1), dims(2));
         std::copy(other.data(), other.data() + other.m_grid.size(), m_grid.data());
-        
+
         title = other.title;
         format = other.format;
         data_type = other.data_type;
@@ -51,22 +54,22 @@ PeriodicGrid PeriodicGrid::load_pgrid(const std::string &filename) {
     if (!file.is_open()) {
         throw std::runtime_error("Could not open pgrid file: " + filename);
     }
-    
+
     PeriodicGrid grid;
     grid.format = GridFormat::PeriodicGrid;
-    
+
     auto header = read_header(file);
     grid.title = std::string(header.title);
     grid.data_type = header.data_type == 0 ? GridDataType::Raw : GridDataType::Indexed;
     grid.cell_parameters = header.cell_params;
-    
+
     if (header.data_type == 1) {
         grid.symmetry = read_symmetry_info(file, header);
         grid.read_indexed_data(file, header, grid.symmetry);
     } else {
         grid.read_raw_data(file, header);
     }
-    
+
     return grid;
 }
 
@@ -75,22 +78,22 @@ PeriodicGrid PeriodicGrid::load_ggrid(const std::string &filename) {
     if (!file.is_open()) {
         throw std::runtime_error("Could not open ggrid file: " + filename);
     }
-    
+
     PeriodicGrid grid;
     grid.format = GridFormat::GeneralGrid;
-    
+
     auto header = read_header(file);
     grid.title = std::string(header.title);
     grid.data_type = header.data_type == 0 ? GridDataType::Raw : GridDataType::Indexed;
     grid.cell_parameters = header.cell_params;
-    
+
     if (header.data_type == 1) {
         grid.symmetry = read_symmetry_info(file, header);
         grid.read_indexed_data(file, header, grid.symmetry);
     } else {
         grid.read_raw_data(file, header);
     }
-    
+
     return grid;
 }
 
@@ -110,7 +113,7 @@ void PeriodicGrid::save_pgrid(const std::string &filename) const {
     if (!file.is_open()) {
         throw std::runtime_error("Could not open file for writing: " + filename);
     }
-    
+
     // Ensure we're saving as periodic
     if (format != GridFormat::PeriodicGrid) {
         PeriodicGrid periodic_copy = *this;
@@ -119,19 +122,19 @@ void PeriodicGrid::save_pgrid(const std::string &filename) const {
         periodic_copy.save_pgrid(filename);
         return;
     }
-    
+
     write_header(file);
-    
+
     if (data_type == GridDataType::Indexed) {
         // Write symmetry info
         int npos = symmetry.num_operations;
         int ncen = symmetry.centrosymmetric;
         int nsub = symmetry.num_sublattices;
[…]
