### 2025-12-04 38c2822
This converts vector definitions from a compiler extension to std::vector, as they don't work with msvc
This converts vector definitions from a compiler extension to std::vector, as they don't work with msvc
