+++ src/3rdparty/libecpint/ecpint.cpp
@@ -1,4 +1,4 @@
-/* 
+/*
  *      Copyright (c) 2020 Robert Shaw
  *		This file is a part of Libecpint.
  *
@@ -35,13 +35,13 @@ namespace libecpint {
 	ECPIntegral::ECPIntegral(int maxLB, int maxLU, int deriv,
 							 double thresh, unsigned smallGrid, unsigned bigGrid) {
 		// Make sure library can perform requested integrals
-		assert(maxLB+deriv <= LIBECPINT_MAX_L); 
+		assert(maxLB+deriv <= LIBECPINT_MAX_L);
 		assert(maxLU <= LIBECPINT_MAX_L);
-		
+
 		// Initialise singletons
 		initFactorials();
 		zero = nonzero = skipped = 0;
-		
+
 		// Initialise angular and radial integrators
 		angInts.init(maxLB + deriv, maxLU);
 		angInts.compute();
@@ -65,7 +65,7 @@ namespace libecpint {
 		for (int x = L; x >= 0; x--) {
 			for (int y = L-x; y >= 0; y--) {
 				z = L - x - y;
-			
+
 				for (int k = 0; k<= x; k++) {
 					Ck = calcC(x, k, A[0]);
 					for (int l = 0; l <= y; l++) {
@@ -85,7 +85,7 @@ namespace libecpint {
 
 		int LA = data.LA; int LB = data.LB;
 		int maxLBasis = data.maxLBasis;
-	
+
 		// Build radial integrals
 		int L = LA + LB;
 		TwoIndex<double> temp;
@@ -96,11 +96,11 @@ namespace libecpint {
 				for (int m = -l; m <= l; m++) radials(ix, l, l+m) = temp(l, l+m);
 			}
 		}
-	
+
 		// Unpack positions
 		double Ax = data.A[0]; double Ay = data.A[1]; double Az = data.A[2];
 		double Bx = data.B[0]; double By = data.B[1]; double Bz = data.B[2];
-	
+
 		// Calculate chi_ab for all ab in shells
 		int z1, z2, lparity, mparity, msign, ix, k, l, m;
 		double C;
@@ -109,19 +109,19 @@ namespace libecpint {
 			for (int y1 = LA-x1; y1 >= 0; y1--) {
 				z1 = LA - x1 - y1;
 				nb = 0;
-			
+
 				for (int x2 = LB; x2 >= 0; x2--) {
 					for (int y2 = LB-x2; y2 >= 0; y2--) {
 						z2 = LB - x2 - y2;
-					
+
 						for (int k1 = 0; k1 <= x1; k1++) {
 							for (int k2 = 0; k2 <= x2; k2++) {
 								k = k1 + k2;
-							
+
 								for (int l1 = 0; l1 <= y1; l1++) {
 									for (int l2 = 0; l2 <= y2; l2++) {
 										l = l1 + l2;
-									
+
 										for (int m1 = 0; m1 <= z1; m1++) {
 											for (int m2 = 0; m2 <= z2; m2++){
 												m = m1 + m2;
@@ -129,18 +129,18 @@ namespace libecpint {
 												if ( fabs(C) > 1e-14 ) {
 													// Build radial integrals
 													ix = k + l + m;
-													
+
 													// Certain terms can be neglected as the angular integrals will always be zero
-													// See Flores06 appendix for details. 
-													lparity = ix % 2; 
+													// See Flores06 appendix for details.
+													lparity = ix % 2;
 													msign = 1 - 2*(l%2);
 													mparity = (lparity + m) % 2;
[…]
