### 2025-12-05 ae27105
Add LGPLv3, fix bug in gradients.h #44
