+++ LICENSE.txt
@@ -1,17 +1,21 @@
-occ, a program for quantum chemistry & crystallography
-Copyright (C) 2023 Peter Spackman
+occ, a library for quantum chemistry & crystallography
+Copyright (C) 2023-2025 Peter Spackman
 
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+This software is dual-licensed. You may use, redistribute, and/or modify
+it under the terms of either:
 
-This program is distributed in the hope that it will be useful,
+  - The GNU General Public License, version 3 or later (GPLv3+), OR
+  - The GNU Lesser General Public License, version 3 or later (LGPLv3+)
+
+at your option.
+
+This software is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU General Public License or GNU Lesser General Public License for
+more details.
 
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
+You should have received a copy of both licenses along with this software.
+If not, see <https://www.gnu.org/licenses/>.
 
 If you are interested in other licensing options, please contact the author.
