### 2025-12-06 #9421 @xuanxu
@rkurchin the branch with the paper ([paper-joss](https://github.com/peterspackman/occ/tree/paper-joss)) does not include the code
