### 2025-12-09 48b8525
Removed unnecessary libraries from integral_engine_df
