### 2025-12-09 fed5cc0
small changes to cint_interface.h
