+++ include/occ/qm/integral_engine.h
@@ -1,12 +1,10 @@
 #pragma once
 #include <occ/core/log.h>
-#include <occ/core/multipole.h>
 #include <occ/core/parallel.h>
 #include <occ/core/timings.h>
 #include <occ/qm/cint_interface.h>
 #include <occ/qm/expectation.h>
 #include <occ/qm/mo.h>
-#include <occ/qm/shell.h>
 #include <occ/qm/shellblock_norm.h>
 #include <optional>
 #include <unsupported/Eigen/CXX11/Tensor>
