### 2025-12-10 a6bfd19
Removed unnecessary libraries
