### 2025-12-10 6b9fb28
Removed unnecessary libraries mostly linear_algebra double includes
