+++ src/main/xdm.cpp
@@ -1,14 +1,11 @@
 #include <cmath>
 #include <filesystem>
 #include <fmt/core.h>
-#include <iostream>
-#include <occ/core/linear_algebra.h>
 #include <occ/core/log.h>
 #include <occ/io/fchkreader.h>
 #include <occ/io/moldenreader.h>
 #include <occ/qm/wavefunction.h>
 #include <occ/xdm/xdm.h>
-#include <vector>
 
 using occ::qm::Wavefunction;
 
