+++ src/elastic_fit/elastic_fitting.cpp
@@ -1,4 +1,3 @@
-#include <algorithm>
 #include <ankerl/unordered_dense.h>
 #include <cmath>
 #include <fmt/os.h>
