+++ src/elastic_fit/elastic_fitting.cpp
@@ -1,5 +1,4 @@
 #include <ankerl/unordered_dense.h>
-#include <cmath>
 #include <fmt/os.h>
 #include <occ/core/log.h>
 #include <occ/core/units.h>
