### 2025-12-12 84338ea
Removed unnecessary libraries from source files
