+++ src/io/moldenreader.cpp
@@ -3,7 +3,6 @@
 #include <occ/core/log.h>
 #include <occ/core/timings.h>
 #include <occ/core/util.h>
-#include <occ/gto/gto.h>
 #include <occ/io/moldenreader.h>
 #include <scn/scan.h>
 
