### 2025-12-12 e6c1810
Removed unnecessary libraries from source files (final)
