+++ src/main/occ_embed.cpp
@@ -1,29 +1,23 @@
-#include <CLI/Config.hpp>
-#include <CLI/Formatter.hpp>
+
 #include <filesystem>
 #include <fmt/core.h>
 #include <fmt/ostream.h>
 #include <fstream>
 #include <occ/core/log.h>
-#include <occ/core/parallel.h>
 #include <occ/core/progress.h>
 #include <occ/core/timings.h>
 #include <occ/core/units.h>
 #include <occ/core/util.h>
-#include <occ/dft/hirshfeld.h>
 #include <occ/driver/single_point.h>
 #include <occ/interaction/wolf.h>
-#include <occ/interaction/pairinteraction.h>
 #include <occ/interaction/coulomb.h>
 #include <occ/interaction/ce_energy_model.h>
 #include <occ/io/load_geometry.h>
 #include <occ/xdm/xdm.h>
 #include <occ/io/occ_input.h>
-#include <occ/main/cli_validators.h>
 #include <occ/main/occ_embed.h>
 #include <occ/main/version.h>
 #include <occ/qm/chelpg.h>
-#include <occ/qm/expectation.h>
 #include <occ/qm/external_potential.h>
 #include <occ/qm/hf.h>
 #include <occ/qm/scf.h>
