+++ src/io/orca_json.cpp
@@ -1,4 +1,3 @@
-#include <fmt/ostream.h>
 #include <nlohmann/json.hpp>
 #include <occ/core/log.h>
 #include <occ/core/timings.h>
