+++ src/solvent/draco.cpp
@@ -3,6 +3,8 @@
 #include <occ/core/units.h>
 #include <occ/solvent/draco.h>
 #include <occ/solvent/parameters.h>
+#include <occ/solvent/smd.h>
+#include <occ/core/molecule.h>
 
 namespace occ::solvent::draco {
 
