+++ src/main/occ_dimers.cpp
@@ -10,7 +10,7 @@
 namespace fs = std::filesystem;
 using occ::crystal::Crystal;
 
-inline Crystal read_crystal(const std::string &filename) {
+inline Crystal dimer_read_crystal(const std::string &filename) {
   occ::io::CifParser parser;
   return parser.parse_crystal_from_file(filename).value();
 }
@@ -40,7 +40,7 @@ void run_dimers_subcommand(const DimerGenerationSettings &settings) {
   std::string filename = settings.crystal_filename;
   std::string basename = fs::path(filename).stem().string();
 
-  Crystal c = read_crystal(filename);
+  Crystal c = dimer_read_crystal(filename);
   occ::log::info("Loaded crystal from {}", filename);
   auto molecules = c.symmetry_unique_molecules();
   auto uc_molecules = c.unit_cell_molecules();
