+++ src/elastic_fit/pes.cpp
@@ -1,6 +1,7 @@
 #include <occ/core/constants.h>
 #include <occ/elastic_fit/monkhorst_pack.h>
 #include <occ/elastic_fit/pes.h>
+#include <iomanip>
 
 namespace occ::elastic_fit {
 
