### 2026-01-07 #9609 @peterspackman
> Dear All,
>
> There is not much more to add from my side. This is a nice quantum chemistry code/library with its own niche, and it does what it claims. The web-based interface is particularly commendable, as it allows the program to run on almost any computer without installing locally, which can be very useful when teaching basic DFT courses/workshops. I can recommend this work for publication in JOSS without any reservations.
>
> I do, however, have a couple of minor suggestions (of course, the author is free to ignore these):
>
>     1. I may have missed it, but the instructions for running the test suite are not present. Some notes on how to run existing tests, and possibly how to add new ones, could be helpful for new contributors.
>
>     2. It might be nice to cite a few academic quantum chemistry codes with similar goals. For example, codes such as PySCF are quite modular and hackable, and could serve as useful reference points.
>
>     3. It would also be helpful to briefly mention what kinds of parallelization are supported.
>
>     4. I suggest to mention about the  `help-all` command in the tutorial to see other available options for the user.
>
>
> Best regards, Murali

Those are all helpful suggestions, and I completely agree that those points and the documentation more broadly need clarification and improvement going forward. I'll do my best to address them all in the near future.
