### 2026-01-13 6ca1d06
Fix gradient bug, clean up build acceleration options
