+++ src/qm/integral_engine_grad1.cpp
@@ -232,7 +232,7 @@ coulomb_kernel_grad(cint::IntegralEnvironment &env, const AOBasis &basis,
   };
 
   occ::timing::start(occ::timing::category::fock);
-  detail::evaluate_four_center_tbb<op, kind>(
+  detail::gradient_evaluate_four_center_tbb<op, kind>(
       f, env, basis, shellpairs, Dnorm, Schwarz, precision);
   occ::timing::stop(occ::timing::category::fock);
 
@@ -337,7 +337,7 @@ JKTriple coulomb_exchange_kernel_grad(IntEnv &env, const AOBasis &basis,
   };
 
   occ::timing::start(occ::timing::category::fock);
-  detail::evaluate_four_center_tbb<op, kind>(
+  detail::gradient_evaluate_four_center_tbb<op, kind>(
       f, env, basis, shellpairs, Dnorm, Schwarz, precision);
   occ::timing::stop(occ::timing::category::fock);
 
