### 2026-01-13 3ac32f8
Improve compile times: remove unnecessary includes, add ccache/lld support (#45)
* testing commit

* Test commit

* test

* test on branch

* added ccache and lld checks

* moved PES::solve_linear_system to .cpp to compile once

* Removed unnecessary includes

* Removed unnecessary libraries from integral_engine_df

* small changes to cint_interface.h

* Removed unnecessary includes from smd_salvation.cpp,occ_cube.cpp,mo.cpp

* Removed unnecessary libraries

* Removed unnecessary libraries mostly linear_algebra double includes

* Removed unnecessary libraries

* Removed unnecessary libraries f
[…]
