+++ src/qm/detail/kernel_traits.h
@@ -2,7 +2,7 @@
 #include <occ/qm/cint_interface.h>
 #include <occ/qm/mo.h>
 #include <occ/qm/shell.h>
-#include <vector>
+
 
 namespace occ::qm::detail {
 
