+++ search/functions_3.js
@@ -5,7 +5,7 @@ var searchData=
   ['data_2',['data',['../classocc_1_1core_1_1ElasticTensor.html#ab3fc715ca830d20d9f9473f4479dfbdf',1,'occ::core::ElasticTensor::data()'],['../classocc_1_1core_1_1ElasticTensor.html#a1322d2d198365fe19faa793b8103c7b4',1,'occ::core::ElasticTensor::data() const'],['../structocc_1_1core_1_1numpy_1_1NumpyArray.html#a1d0567415471326db09f5e9010a9db68',1,'occ::core::numpy::NumpyArray::data()'],['../structocc_1_1core_1_1numpy_1_1NumpyArray.html#adb3db88749cccf905c3361079bbe4c9f',1,'occ::core::numpy::NumpyArray::data() const'],['../classocc_1_1geometry_1_1VolumeGrid.html#a63d6e3607db736b090efc7f7bca15f28',1,'occ::geometry::VolumeGrid::data()'],['../classocc_1_1geometry_1_1VolumeGrid.html#a915436084e50c43fb011e54c1a36d316',1,'occ::geometry::VolumeGrid::data() const'],['../classocc_1_1io_1_1Cube.html#a5cd3a0c6b9034b3b0ce99725b6a7ee8d',1,'occ::io::Cube::data() const'],['../classocc_1_1io_1_1Cube.html#a337868a5b715d1c379028722db54cdc6',1,'occ::io::Cube::data()'],['../classocc_1_1io_1_1PeriodicGrid.html#a0f1d01acc45d1699084628740a96b5cc',1,'occ::io::PeriodicGrid::data() const'],['../classocc_1_1io_1_1PeriodicGrid.html#ad1a532270999f6e082980a7b8bafdb0a',1,'occ::io::PeriodicGrid::data()']]],
   ['default_5fproperties_3',['default_properties',['../namespaceocc_1_1isosurface.html#ab16fc330a28219f949bbbe9bed9fb603',1,'occ::isosurface']]],
   ['defaulteps_4',['defaultEps',['../namespacequickhull.html#aed54901ccc77c377d9d2f94ce6a0888c',1,'quickhull']]],
-  ['deformationdensityfunctor_5',['deformationdensityfunctor',['../structocc_1_1main_1_1DeformationDensityFunctor.html#a5301e50a4281ba0bad0686c58694b3fc',1,'occ::main::DeformationDensityFunctor::DeformationDensityFunctor()'],['../classocc_1_1isosurface_1_1DeformationDensityFunctor.html#a2d12fdddf89bde035897c188804bd0d2',1,'occ::isosurface::DeformationDensityFunctor::DeformationDensityFunctor(const Wavefunction &amp;wfn, SpinConstraint=SpinConstraint::Total)'],['../classocc_1_1isosurface_1_1DeformationDensityFunctor.html#add646be9c1b9441a973b576552e64bf3',1,'occ::isosurface::DeformationDensityFunctor::DeformationDensityFunctor(const occ::core::Molecule &amp;mol, const occ::qm::Wavefunction &amp;wfn, const occ::slater::InterpolatorParams &amp;params={})']]],
+  ['deformationdensityfunctor_5',['deformationdensityfunctor',['../structocc_1_1main_1_1DeformationDensityFunctor.html#a5301e50a4281ba0bad0686c58694b3fc',1,'occ::main::DeformationDensityFunctor::DeformationDensityFunctor()'],['../classocc_1_1isosurface_1_1DeformationDensityFunctor.html#add646be9c1b9441a973b576552e64bf3',1,'occ::isosurface::DeformationDensityFunctor::DeformationDensityFunctor()']]],
[…]
