### 2026-01-20 #9609 @peterspackman
> Hi All,
>
> I have completed reviewing this. It is certainly an impressive piece of software!. I was able to compile, install, and run everything very easily which is impressive. I checked correctness against another DFT package I am familiar with.
>
> My only minor suggestions relate to the documentation, it is rather fragmented and hard to find, e.g:
>
>     * the main website and web interface: https://getocc.xyz/
>
>     * c++ api: https://peterspackman.github.io/occ/
>
>     * python interface: https://github.com/peterspackman/occ#python-api-examples
>
>     * a Wiki: https://github.com/peterspackman/occ/wiki
>
>
> and the paper does not mention other open source packages such as psi4 or pyscf.
>
> But these are minor points and should not block submission.

Glad to hear it all (mostly) worked - and updating, centralising and improving the documentation is a priority - the main website (getocc.xyz) was a first start for doing that in an improved way for end-users.

I'm totally happy to add some discussion/mention of other open source quantum chemistry programs/libraries if that's appropriate to the paper.
