+++ 3rdparty/CMakeLists.txt
@@ -152,6 +152,10 @@ CPMAddPackage(
         ${LIBCINT_BUILD_TYPE}
     VERSION "6.1.2"
 )
+if (libcint_ADDED AND NOT NIX_BUILD)
+    file(COPY_FILE ${CMAKE_CURRENT_SOURCE_DIR}/../cmake/FindQUADMATH.cmake ${libcint_SOURCE_DIR}/cmake/FindQUADMATH.cmake)
+    message(WARNING "ADDED LIBCINT")
+endif ()
 
 # Apply the flags to the target after it's created
 if(TARGET cint)
