+++ CMakeLists.txt
@@ -131,6 +131,21 @@ endif()
 
 include("${CMAKE_CURRENT_SOURCE_DIR}/cmake/CPM.cmake")
 # don't use install commands from third party
+if (NIX_BUILD)
+find_package(fmt REQUIRED)
+find_package(spdlog REQUIRED)
+add_library(spdlog::spdlog ALIAS spdlog::spdlog_header_only)
+find_package(TBB REQUIRED)
+find_package(scn REQUIRED)
+find_package(unordered_dense REQUIRED)
+find_package(CLI11 REQUIRED)
+find_package(nlohmann_json REQUIRED)
+find_package(gemmi REQUIRED)
+add_library(gemmi::gemmi ALIAS gemmi::gemmi_cpp)
+find_package(dftd4 REQUIRED)
+find_package(lbfgspp REQUIRED)
+add_library(LBFGSpp::LBFGSpp ALIAS lbfgspp)
+endif ()
 add_subdirectory(3rdparty EXCLUDE_FROM_ALL)
 add_subdirectory(include)
 add_subdirectory(src)
