+++ CMakeLists.txt
@@ -145,6 +145,8 @@ add_library(gemmi::gemmi ALIAS gemmi::gemmi_cpp)
 find_package(dftd4 REQUIRED)
 find_package(lbfgspp REQUIRED)
 add_library(LBFGSpp::LBFGSpp ALIAS lbfgspp)
+find_package(libcint REQUIRED)
+find_package(Libxc REQUIRED)
 endif ()
 add_subdirectory(3rdparty EXCLUDE_FROM_ALL)
 add_subdirectory(include)
