### 2026-01-26 0ddca35
Add COSX seminumerical exchange, Split-RI-J Coulomb, and auto auxiliary basis
Adds McMurchie-Davidson based ESP integrals, Hermite-based 3-center integral
kernels, and pivoted Cholesky automatic auxiliary basis generation.
