+++ src/main/properties.cpp
@@ -4,6 +4,7 @@
 #include <fstream>
 #include <occ/core/log.h>
 #include <occ/core/units.h>
+#include <occ/numint/molecular_grid.h>
 #include <occ/main/properties.h>
 #include <occ/qm/chelpg.h>
 #include <occ/qm/hf.h>
