+++ include/occ/driver/crystal_growth.h
@@ -5,6 +5,7 @@
 #include <occ/cg/solvation_contribution.h>
 #include <occ/cg/solvent_surface.h>
 #include <occ/core/dimer.h>
+#include <occ/core/log.h>
 #include <occ/core/molecule.h>
 #include <occ/core/point_group.h>
 #include <occ/core/timings.h>
@@ -12,7 +13,8 @@
 #include <occ/driver/single_point.h>
 #include <occ/interaction/pair_energy.h>
 #include <occ/interaction/pairinteraction.h>
-#include <occ/io/wavefunction_json.h>
+#include <occ/qm/io/wavefunction_json.h>
+#include <occ/qm/wavefunction.h>
 
 namespace occ::driver {
 

[…]
