+++ structocc_1_1io_1_1GridSettings.js
@@ -1,7 +1,11 @@
 var structocc_1_1io_1_1GridSettings =
 [
+    [ "has_angular_regions", "structocc_1_1io_1_1GridSettings.html#a1ca1e45140539c3c225f44b5bae78378", null ],
+    [ "int_acc_string", "structocc_1_1io_1_1GridSettings.html#aa7e2228991179c023b8b891d144f5188", null ],
     [ "operator!=", "structocc_1_1io_1_1GridSettings.html#a979222d60c32cfa18e2a7dfeb1cdb51d", null ],
     [ "operator==", "structocc_1_1io_1_1GridSettings.html#ad1629f756b11f97ca33aa1fdf0b64f39", null ],
+    [ "angular_regions", "structocc_1_1io_1_1GridSettings.html#a2c91a49afbde53c980ac87f9ce1a144f", null ],
+    [ "int_acc", "structocc_1_1io_1_1GridSettings.html#adeb82c01c4dbbe1db7d21b3e272b81f1", null ],
     [ "max_angular_points", "structocc_1_1io_1_1GridSettings.html#afe9e84413557f08fcd8a931628cb5d6e", null ],
     [ "min_angular_points", "structocc_1_1io_1_1GridSettings.html#a418d0f9e894467350cbc64ffbdca598b", null ],
     [ "pruning_scheme", "structocc_1_1io_1_1GridSettings.html#a43adfd375523983981a74c981f270f82", null ],
