+++ search/all_d.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['m_0',['m',['../structocc_1_1gto_1_1MomentaSpherical.html#a8a9b58d31cc767f52e3c86b89a382430',1,'occ::gto::MomentaSpherical::m'],['../structocc_1_1gto_1_1Momenta.html#af0edf1d8bba9d9a449cc98d32f86d3bb',1,'occ::gto::Momenta::m']]],
+  ['m_0',['m',['../structocc_1_1gto_1_1Momenta.html#af0edf1d8bba9d9a449cc98d32f86d3bb',1,'occ::gto::Momenta::m'],['../structocc_1_1gto_1_1MomentaSpherical.html#a8a9b58d31cc767f52e3c86b89a382430',1,'occ::gto::MomentaSpherical::m']]],
   ['m_5fao_5fengine_1',['m_ao_engine',['../classocc_1_1qm_1_1PostHFMethod.html#a8f735f35aee3c339efdd37e03576f9ec',1,'occ::qm::PostHFMethod']]],
   ['m_5fatoms_2',['m_atoms',['../classocc_1_1qm_1_1SCFMethodBase.html#ad3995211c743b6e3400ff9cc2f32d4a3',1,'occ::qm::SCFMethodBase']]],
   ['m_5fcharge_3',['m_charge',['../classocc_1_1qm_1_1SCFMethodBase.html#ab87e63b8c59b5d85ddf807bb1d52cdfd',1,'occ::qm::SCFMethodBase']]],
@@ -35,7 +35,7 @@ var searchData=
   ['m_5fmostdistantpoint_32',['m_mostDistantPoint',['../structquickhull_1_1MeshBuilder_1_1Face.html#ab4716d5bd1dc131e8ff9d08f9a14117f',1,'quickhull::MeshBuilder::Face']]],
   ['m_5fmostdistantpointdist_33',['m_mostDistantPointDist',['../structquickhull_1_1MeshBuilder_1_1Face.html#aaa4d93609ff57160e813e6bebcd15ddd',1,'quickhull::MeshBuilder::Face']]],
   ['m_5fnearest_5fdimers_34',['m_nearest_dimers',['../classocc_1_1driver_1_1CrystalGrowthCalculator.html#a610b893c63368ecd5d951fdebc23c412',1,'occ::driver::CrystalGrowthCalculator']]],
-  ['m_5fnext_35',['m_next',['../structquickhull_1_1HalfEdgeMesh_1_1HalfEdge.html#a9e4c6c2c4eb56e4219a88022c7826846',1,'quickhull::HalfEdgeMesh::HalfEdge::m_next'],['../structquickhull_1_1MeshBuilder_1_1HalfEdge.html#afaf06fb0b72d82f34185f7297f6dd6ea',1,'quickhull::MeshBuilder::HalfEdge::m_next']]],
+  ['m_5fnext_35',['m_next',['../structquickhull_1_1MeshBuilder_1_1HalfEdge.html#afaf06fb0b72d82f34185f7297f6dd6ea',1,'quickhull::MeshBuilder::HalfEdge::m_next'],['../structquickhull_1_1HalfEdgeMesh_1_1HalfEdge.html#a9e4c6c2c4eb56e4219a88022c7826846',1,'quickhull::HalfEdgeMesh::HalfEdge::m_next']]],
   ['m_5fnum_5felectrons_36',['m_num_electrons',['../classocc_1_1qm_1_1SCFMethodBase.html#ae87859b6fb8b1891f0d9ac5146a9d1a8',1,'occ::qm::SCFMethodBase']]],
[…]
