+++ search/classes_9.js
@@ -2,8 +2,8 @@ var searchData=
 [
   ['jkpair_0',['JKPair',['../structocc_1_1qm_1_1JKPair.html',1,'occ::qm']]],
   ['jktriple_1',['JKTriple',['../structocc_1_1qm_1_1JKTriple.html',1,'occ::qm']]],
-  ['jsonbasis_2',['JsonBasis',['../structocc_1_1io_1_1JsonBasis.html',1,'occ::io']]],
-  ['jsonbasisreader_3',['JsonBasisReader',['../structocc_1_1io_1_1JsonBasisReader.html',1,'occ::io']]],
+  ['jsonbasis_2',['JsonBasis',['../structocc_1_1gto_1_1io_1_1JsonBasis.html',1,'occ::gto::io']]],
+  ['jsonbasisreader_3',['JsonBasisReader',['../structocc_1_1gto_1_1io_1_1JsonBasisReader.html',1,'occ::gto::io']]],
   ['jsonwavefunctionreader_4',['JsonWavefunctionReader',['../structocc_1_1io_1_1JsonWavefunctionReader.html',1,'occ::io']]],
   ['jsonwavefunctionwriter_5',['JsonWavefunctionWriter',['../structocc_1_1io_1_1JsonWavefunctionWriter.html',1,'occ::io']]]
 ];

[…]
