### 2026-01-26 a6c0101
Add multipole S-functions module for electrostatic interactions
Introduces the occ_mults library implementing Orient-compatible multipole
S-functions for computing electrostatic interaction energies, forces, and
torques between rigid molecular pairs. Includes:

- S-function evaluators for multipole ranks 0-5 (charge through hexadecapole)
- Rigid body dynamics with angle-axis orientation parameterization
- Derivative transforms, coordinate systems, and short-range potentials
