+++ src/mults/derivative_transform.cpp
@@ -1,0 +1,712 @@
+#include <occ/mults/derivative_transform.h>
+#include <stdexcept>
+#include <fmt/core.h>
+#include <fmt/format.h>
+
+namespace occ::mults {
+
+Vec3 DerivativeTransform::cross(const Vec3& a, const Vec3& b) {
+    Vec3 result;
+    result[0] = a[1] * b[2] - a[2] * b[1];
+    result[1] = a[2] * b[0] - a[0] * b[2];
+    result[2] = a[0] * b[1] - a[1] * b[0];
+    return result;
+}
+
+Mat DerivativeTransform::compute_D1_angle_axis(
+    const CoordinateSystem& coords,
+    const Mat3& M_A,
+    const Mat3& M_B,
+    const std::array<Mat3, 3>& M1_A,
+    const std::array<Mat3, 3>& M1_B,
+    const Vec3& a,
+    const Vec3& b) {
+
+    // Initialize D1 matrix [16 x 12]
+    Mat D1 = Mat::Zero(NUM_INTERMEDIATE_VARS, NUM_EXTERNAL_COORDS);
+
+    const double r = coords.r;
+    if (r < 1e-15) {
+        throw std::runtime_error("Distance too small for derivative calculation");
+    }
+
+    // Extract coordinate components (following Orient's pairinfo_aa setup)
+    const Vec3& rab = coords.rab;  // rb - ra
+    const Vec3& er = coords.er;    // unit vector
+
+    // CRITICAL: For angle-axis parametrization, we use BODY-FRAME-PROJECTED geometry:
+    //   e1r = M_A^T · er  (unit vector in A's body frame)
+    //   e2r = -M_B^T · er (unit vector in B's body frame, with sign flip)
+    //
+    // These unit vectors DO depend on molecular rotations through M_A and M_B,
+    // so their derivatives w.r.t. angle-axis parameters are NON-ZERO.
+    //
+    // For point multipoles: srotlink = I (no site rotation relative to COM)
+    //                       slink = 0 (sites at COM)
+    // So: se_A = M_A, se_B = M_B (body frame orientations)
+    Mat3 se_A = M_A;  // Site orientation = molecular orientation
+    Mat3 se_B = M_B;  // Site orientation = molecular orientation
+
+    // Unit vectors in body frames (from CoordinateSystem, precomputed)
+    // coords should have been created with from_body_frame(ra, rb, M_A, M_B)
+    Vec3 e1r = coords.raxyz();    // e1r_body from coords
+    Vec3 e2r = coords.rbxyz();    // e2r_body from coords
+
+    // Helper vectors (Orient line 358-359)
+    Vec3 ra_vec = rab + a;  // From A's COM to site B
+    Vec3 rb_vec = rab - b;  // From site A to B's COM
+
+    // ========== TRANSLATIONAL DERIVATIVES (columns 0-2, 6-8) ==========
+    // Following Orient lines 383-394
+
+    // Distance r (row 15) - Orient line 384-385
+    D1(15, 0) = -rab[0] / r;
+    D1(15, 1) = -rab[1] / r;
+    D1(15, 2) = -rab[2] / r;
+    D1(15, 6) = rab[0] / r;
+    D1(15, 7) = rab[1] / r;
+    D1(15, 8) = rab[2] / r;
+
+    // R·wa (rows 0-2) - Orient line 387-388
+    // ∂(rab · M_A)/∂x_A = -M_A^T, ∂(rab · M_A)/∂x_B = +M_A^T
[…]
