+++ tests/coordinate_system_tests.cpp
@@ -1,0 +1,207 @@
+#include <catch2/catch_test_macros.hpp>
+#include <catch2/catch_approx.hpp>
+#include <occ/mults/coordinate_system.h>
+#include <cmath>
+
+using namespace occ::mults;
+using occ::Vec3;
+using Catch::Approx;
+
+TEST_CASE("CoordinateSystem - basic functionality", "[mults][coordinates]") {
+    
+    SECTION("Simple axis-aligned case") {
+        Vec3 ra(0.0, 0.0, 0.0);
+        Vec3 rb(1.0, 0.0, 0.0);  // Unit vector along x-axis
+        auto cs = CoordinateSystem::from_points(ra, rb);
+        
+        // Basic calculations
+        REQUIRE(cs.r == Approx(1.0));
+        REQUIRE(cs.dx() == Approx(1.0));
+        REQUIRE(cs.dy() == Approx(0.0));
+        REQUIRE(cs.dz() == Approx(0.0));
+        
+        // Unit vector should be (1, 0, 0)
+        REQUIRE(cs.er[0] == Approx(1.0));
+        REQUIRE(cs.er[1] == Approx(0.0));
+        REQUIRE(cs.er[2] == Approx(0.0));
+        
+        // Orient convention: rax = -er, rbx = +er
+        REQUIRE(cs.rax() == Approx(-1.0));
+        REQUIRE(cs.ray() == Approx(0.0));
+        REQUIRE(cs.raz() == Approx(0.0));
+        REQUIRE(cs.rbx() == Approx(1.0));
+        REQUIRE(cs.rby() == Approx(0.0));
+        REQUIRE(cs.rbz() == Approx(0.0));
+        
+        // Validation
+        REQUIRE(cs.is_valid());
+    }
+    
+    SECTION("Diagonal case - matches Orient test coordinates") {
+        // This matches our Orient validation case
+        Vec3 ra(-2.0, -2.0, -2.0);
+        Vec3 rb(2.0, 2.0, 2.0);
+        auto cs = CoordinateSystem::from_points(ra, rb);
+        
+        // Distance should be sqrt((4)² + (4)² + (4)²) = sqrt(48) = 4*sqrt(3)
+        double expected_r = 4.0 * std::sqrt(3.0);
+        REQUIRE(cs.r == Approx(expected_r));
+        
+        // Unit vector components should all be 1/sqrt(3) ≈ 0.57735...
+        double expected_unit = 1.0 / std::sqrt(3.0);
+        REQUIRE(cs.er[0] == Approx(expected_unit));
+        REQUIRE(cs.er[1] == Approx(expected_unit));
+        REQUIRE(cs.er[2] == Approx(expected_unit));
+        
+        // Orient coordinates should match our known values
+        REQUIRE(cs.rax() == Approx(-expected_unit));
+        REQUIRE(cs.ray() == Approx(-expected_unit));
+        REQUIRE(cs.raz() == Approx(-expected_unit));
+        REQUIRE(cs.rbx() == Approx(expected_unit));
+        REQUIRE(cs.rby() == Approx(expected_unit));
+        REQUIRE(cs.rbz() == Approx(expected_unit));
+        
+        REQUIRE(cs.is_valid());
+    }
+    
+    SECTION("2D case in xy-plane") {
+        Vec3 ra(-1.0, -1.0, 0.0);
+        Vec3 rb(1.0, 1.0, 0.0);
+        auto cs = CoordinateSystem::from_points(ra, rb);
+        
+        // Distance should be sqrt(4 + 4) = 2*sqrt(2)
+        double expected_r = 2.0 * std::sqrt(2.0);
+        REQUIRE(cs.r == Approx(expected_r));
+        
[…]
