### 2026-01-26 #9609 @rkurchin
Great, thanks all – I'm going to swap over to my AEiC hat and start on some final checks for this. @peterspackman, can you please do the following:
* take a look at the missing/skip DOI's above and see if any can be fleshed out (often the automation doesn't pick up suggested citations for software, so you may have to poke around some documentation/repos for those skip ones)
* If you haven't already, issue a new tagged release with all the code/docs changes from the review and send me the version tag here
* Generate a DOI (e.g. via Zenodo) for that same version of the software; make sure the deposit title, author list, and license match your paper and repository respectively, and send the DOI here
