### 2026-01-26 #9609 @rkurchin
In the meantime, here are some copyediting comments...

### Main text
- [ ] line 20: missing space before citation
- [ ] first two paragraphs of "Implementation and Features": a bunch more of the same, make sure there are spaces before the citations
- [ ] suggest capitalizing both Coulomb (line 59) and Gaussians (line 106)
- [ ] again in first paragraph of Acknowledgements

### References
Throughout, check that you've `{Enforced}` appropriate capitalization in your `.bib` for things like London, C++, Python, Hartree-Fock, van der Waals, Gaussian, etc.
