### 2026-01-27 #9609 @peterspackman
> Great, thanks all – I'm going to swap over to my AEiC hat and start on some final checks for this. [@peterspackman](https://github.com/peterspackman), can you please do the following:
>
>     * take a look at the missing/skip DOI's above and see if any can be fleshed out (often the automation doesn't pick up suggested citations for software, so you may have to poke around some documentation/repos for those skip ones)
>
>     * If you haven't already, issue a new tagged release with all the code/docs changes from the review and send me the version tag here
>
>     * Generate a DOI (e.g. via Zenodo) for that same version of the software; make sure the deposit title, author list, and license match your paper and repository respectively, and send the DOI here

Great, I'll go through and update the paper! Based on the comments in the review I think the documentation updates are not a blocker and there were no required code changes so I'll get onto it as soon as possible. I think in the interests of getting this done it's not worth waiting finalising this paper for me to unify all the documentation changes (that might take a while) but I greatly appreciate the suggestions and will get onto it.

Hopefully I'll have this done this all done this afternoon.
