### 2026-01-27 #9609 @peterspackman
Ok I believe I've done what's required, the tagged release is v0.8.8-paper and the DOI is available on zenodo here: https://doi.org/10.5281/zenodo.18383210 with the title + author list matching the paper.
