### 2026-01-27 f6524fa
Address JOSS review copyediting and reference fixes
