+++ paper.bib
@@ -1,15 +1,16 @@
 @article{Lehtola2018,
-  title={Recent developments in libxc: A comprehensive library of functionals for density functional theory},
+  title={Recent developments in {libxc}: A comprehensive library of functionals for density functional theory},
   author={Lehtola, Susi and Steigemann, Conrad and Oliveira, Micael JT and Marques, Miguel AL},
   journal={SoftwareX},
   volume={7},
   pages={1--5},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2017.11.002}
 }
 
 @article{Mackenzie2017,
-  title={CrystalExplorer model energies and energy frameworks: extension to metal coordination compounds, organic salts, solvates and open-shell systems},
+  title={{CrystalExplorer} model energies and energy frameworks: extension to metal coordination compounds, organic salts, solvates and open-shell systems},
   author={Mackenzie, Campbell F and Spackman, Peter R and Jayatilaka, Dylan and Spackman, Mark A},
   journal={IUCrJ},
   volume={4},
@@ -48,7 +49,7 @@ url = {https://doi.org/10.1107/S1600576721002910},
 }
 
 @article{Klamt1993,
-  title = {COSMO: a new approach to dielectric screening in solvents with explicit expressions for the screening energy and its gradient},
+  title = {{COSMO}: a new approach to dielectric screening in solvents with explicit expressions for the screening energy and its gradient},
   ISSN = {1364-5471},
   url = {http://dx.doi.org/10.1039/P29930000799},
   DOI = {10.1039/p29930000799},
@@ -76,7 +77,7 @@ url = {https://doi.org/10.1107/S1600576721002910},
 }
 
 @article{Johnson2006,
-  title = {A post-Hartree-Fock model of intermolecular interactions: Inclusion of higher-order corrections},
+  title = {A post-{Hartree-Fock} model of intermolecular interactions: Inclusion of higher-order corrections},
   volume = {124},
   ISSN = {1089-7690},
   url = {http://dx.doi.org/10.1063/1.2190220},
@@ -104,7 +105,7 @@ url = {https://doi.org/10.1107/S1600576721002910},
 }
 
 @article{OterodelaRoza2012,
-  title = {Van der Waals interactions in solids using the exchange-hole dipole moment model},
+  title = {{Van der Waals} interactions in solids using the exchange-hole dipole moment model},
   volume = {136},
   ISSN = {1089-7690},
   url = {http://dx.doi.org/10.1063/1.4705760},
@@ -119,7 +120,7 @@ url = {https://doi.org/10.1107/S1600576721002910},
 
 
 @article{Caldeweyher2017,
-  title = {Extension of the D3 dispersion coefficient model},
+  title = {Extension of the {D3} dispersion coefficient model},
   volume = {147},
   ISSN = {1089-7690},
   url = {http://dx.doi.org/10.1063/1.4993215},
@@ -133,7 +134,7 @@ url = {https://doi.org/10.1107/S1600576721002910},
 }
 
 @article{Caldeweyher2019,
-  title = {A generally applicable atomic-charge dependent London dispersion correction},
+  title = {A generally applicable atomic-charge dependent {London} dispersion correction},
   volume = {150},
   ISSN = {1089-7690},
   url = {http://dx.doi.org/10.1063/1.5090222},
[…]
