### 2026-01-27 #9609 @rkurchin
Congratulations @peterspackman and thanks especially to reviewers @muralidhar-nalabothula and @sef43, JOSS couldn't exist without your efforts!
